feat: Organize models package into logical domains
- Split types into core, entities, and game-state domains
- Split schemas into core, entities, and game-state domains
- Maintain same public API with re-exports from main files
- Better separation of concerns for maintainability
- Clear boundaries between Core/Entities/Game State
- Improved team collaboration potential
- All tests still pass (14/14)
- Same import API for consumers
