+++ packages/models/src/schemas/index.ts
@@ -1,0 +1,12 @@
+﻿/**
+ * Schema exports - organized by domain for better maintainability
+ */
+
+// Core schemas
+export * from "./core";
+
+// Entity schemas
+export * from "./entities";
+
+// Game state schemas
+export * from "./game-state";
