+++ packages/models/src/types/index.ts
@@ -1,0 +1,12 @@
+﻿/**
+ * Types exports - organized by domain for better maintainability
+ */
+
+// Core types
+export * from "./core";
+
+// Entity types
+export * from "./entities";
+
+// Game state types
+export * from "./game-state";
